员工端bug修复
1、修复了自己选择自己为审批人、核稿人、分配对象、签发人【更改至不得选择】
2、修复了工作日程，以及含有开始时间和结束时间的申请，未选择开始时间就点击结束时间出现的闪退
3、修复了公文拟稿保存出现长时间的转圈事件
4、修复了新建申请、新建工作日程、公文拟稿时保存以后停留在新建页面导致易出现多次点击保存事件
5、修复了转正申请：查看详情，点击审批进度，显示的审批意见的意见内容显示的是试用期总结内容问题
6、修复了新建招聘申请，薪资标准没有设置可以6k-8k这样子的输入，只能输入数字的问题
7、修复了新建加班申请，添加多个审批人，查看该申请详情时，部分审批人没有显示

+++ app/src/main/java/com/oliveoa/view/myapplication/SelectPersonApprovingActivity.java
@@ -22,6 +22,7 @@ import com.oliveoa.greendao.DepartmentInfoDao;
 import com.oliveoa.greendao.JobTransferApplicationDao;
 import com.oliveoa.greendao.MeetingApplicationDao;
 import com.oliveoa.greendao.OfficialDocumentDao;
+import com.oliveoa.greendao.UserInfoDao;
 import com.oliveoa.greendao.WorkDetailDao;
 import com.oliveoa.jsonbean.ContactJsonBean;
 import com.oliveoa.pojo.ApproveNumber;
@@ -32,6 +33,7 @@ import com.oliveoa.pojo.Item;
 import com.oliveoa.pojo.JobTransferApplication;
 import com.oliveoa.pojo.MeetingApplication;
 import com.oliveoa.pojo.OfficialDocument;
+import com.oliveoa.pojo.UserInfo;
 import com.oliveoa.pojo.WorkDetail;
 import com.oliveoa.util.EntityManager;
 import com.oliveoa.view.R;
@@ -76,6 +78,7 @@ public class SelectPersonApprovingActivity extends AppCompatActivity {
     private JobTransferApplicationDao jobTransferApplicationDao;
     private WorkDetailDao workDetailDao;
     private OfficialDocumentDao officialDocumentDao;
+    private UserInfoDao userInfoDao;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -87,6 +90,7 @@ public class SelectPersonApprovingActivity extends AppCompatActivity {
         initData();
     }
     public void initData(){
+        userInfoDao = EntityManager.getInstance().getUserInfoDao();
         employeeInfoDao = EntityManager.getInstance().getContactInfo();
         departmentInfoDao =EntityManager.getInstance().getDepartmentInfo();
         jobTransferApplicationDao = EntityManager.getInstance().getJobTransferApplicationDao();
@@ -154,14 +158,22 @@ public class SelectPersonApprovingActivity extends AppCompatActivity {
                     Log.i(TAG, "被点击的员工Eid：" + iData.get(groupPosition).get(childPosition).getiEid());
                     mid = iData.get(groupPosition).get(childPosition).getiEid();
                     if (index == 0) {
-                        MeetingApplication application = new MeetingApplication();
-                        application = meetingApplicationDao.queryBuilder().unique();
-                        Log.e(TAG,application.toString());
-                        if (application != null) {
-                            application.setAeid(mid);
-                            meetingApplicationDao.deleteAll();
-                            meetingApplicationDao.insert(application);
-                            back();
+                        UserInfo userInfo = userInfoDao.queryBuilder().unique();
+                        if(userInfo!=null) {
+                            Log.i(TAG,userInfo.toString());
+                            if(!mid.equals(userInfo.getEid())) {
+                                MeetingApplication application = new MeetingApplication();
+                                application = meetingApplicationDao.queryBuilder().unique();
+                                Log.e(TAG, application.toString());
+                                if (application != null) {
+                                    application.setAeid(mid);
+                                    meetingApplicationDao.deleteAll();
+                                    meetingApplicationDao.insert(application);
+                                    back();
+                                }
+                            }else{
+                                Toast.makeText(getApplicationContext(), "提示：审批人不得选自己，请选择其他审批人！", Toast.LENGTH_SHORT).show();
+                            }
                         }
                     }else if(index==10){
                         JobTransferApplication jobTransferApplication = new JobTransferApplication();
@@ -173,49 +185,100 @@ public class SelectPersonApprovingActivity extends AppCompatActivity {
                             back();
                         }
                     }else if(index==12){
-                        WorkDetail workDetail = new WorkDetail();
-                        workDetail = workDetailDao.queryBuilder().unique();
-                        if(workDetail!=null){
-                            workDetail.setAeid(mid);
-                            workDetailDao.deleteAll();
-                            workDetailDao.insert(workDetail);
-                            Log.e(TAG,workDetail.toString());
-                            back();
+                        UserInfo userInfo = userInfoDao.queryBuilder().unique();
+                        if(userInfo!=null) {
+                            Log.i(TAG, userInfo.toString());
+                            if (!mid.equals(userInfo.getEid())) {
+                                WorkDetail workDetail = new WorkDetail();
+                                workDetail = workDetailDao.queryBuilder().unique();
+                                if (workDetail != null) {
+                                    workDetail.setAeid(mid);
+                                    workDetailDao.deleteAll();
+                                    workDetailDao.insert(workDetail);
+                                    Log.e(TAG, workDetail.toString());
+                                    back();
+                                }
+                            }else{
+                                Toast.makeText(getApplicationContext(), "提示：审批人不得选自己，请选择其他审批人！", Toast.LENGTH_SHORT).show();
+                            }
                         }
                     }else if(index==13){
-                       approveNumberDao.deleteAll();
-                       ApproveNumber ap = new ApproveNumber();
-                       ap.setId(mid);
-                       approveNumberDao.insert(ap);
-                       back();
+                        UserInfo userInfo = userInfoDao.queryBuilder().unique();
+                        if(userInfo!=null) {
+                            Log.i(TAG, userInfo.toString());
+                            if (!mid.equals(userInfo.getEid())) {
+                                approveNumberDao.deleteAll();
+                                ApproveNumber ap = new ApproveNumber();
+                                ap.setId(mid);
+                                approveNumberDao.insert(ap);
+                                back();
+                            }else {
+                                Toast.makeText(getApplicationContext(), "提示：工作分配对象不得选自己，请选择其他审批人！", Toast.LENGTH_SHORT).show();
+                            }
+                         }
                     }else if(index==14){
-                        OfficialDocument officialDocument = new OfficialDocument();
-                        officialDocument = officialDocumentDao.queryBuilder().unique();
-                        if(officialDocument!=null){
-                            officialDocument.setNuclearDraftEid(mid);
-                            officialDocumentDao.deleteAll();
-                            officialDocumentDao.insert(officialDocument);
-                            Log.e(TAG,officialDocument.toString());
-                            back();
+                        UserInfo userInfo = userInfoDao.queryBuilder().unique();
+                        if(userInfo!=null) {
+                            Log.i(TAG, userInfo.toString());
+                            if (!mid.equals(userInfo.getEid())) {
+                                OfficialDocument officialDocument = new OfficialDocument();
+                                officialDocument = officialDocumentDao.queryBuilder().unique();
+                                if (officialDocument != null) {
+                                    officialDocument.setNuclearDraftEid(mid);
+                                    officialDocumentDao.deleteAll();
+                                    officialDocumentDao.insert(officialDocument);
+                                    Log.e(TAG, officialDocument.toString());
+                                    back();
+                                }
+                            }else{
+                                Toast.makeText(getApplicationContext(), "提示：核稿人不得选自己，请选择其他审批人！", Toast.LENGTH_SHORT).show();
+                            }
                         }
                     }else if(index==15){
-                        OfficialDocument officialDocument = new OfficialDocument();
-                        officialDocument = officialDocumentDao.queryBuilder().unique();
-                        if(officialDocument!=null){
-                            officialDocument.setIssuedEid(mid);
-                            officialDocumentDao.deleteAll();
-                            officialDocumentDao.insert(officialDocument);
-                            Log.e(TAG,officialDocument.toString());
-                            back();
+                        UserInfo userInfo = userInfoDao.queryBuilder().unique();
+                        if(userInfo!=null) {
+                            Log.i(TAG, userInfo.toString());
+                            if (!mid.equals(userInfo.getEid())) {
+                                OfficialDocument officialDocument = new OfficialDocument();
+                                officialDocument = officialDocumentDao.queryBuilder().unique();
+                                if (officialDocument != null) {
+                                    officialDocument.setIssuedEid(mid);
+                                    officialDocumentDao.deleteAll();
+                                    officialDocumentDao.insert(officialDocument);
+                                    Log.e(TAG, officialDocument.toString());
+                                    back();
+                                }
+                            }else{
+                                Toast.makeText(getApplicationContext(), "提示：签发人不得选自己，请选择其他审批人！", Toast.LENGTH_SHORT).show();
+                            }
                         }
                     }else {
+
                         ApproveNumber ap = approveNumberDao.queryBuilder().where(ApproveNumberDao.Properties.Id.eq(mid)).unique();
                         if (ap == null) {
-                            approveNumber.setId(mid);
-                            approveNumberDao.insert(approveNumber);
-                            back();
+                            if(index==4) {
+                                approveNumber.setId(mid);
+                                approveNumberDao.insert(approveNumber);
+                                back();
+                            }else{
+                                UserInfo userInfo = userInfoDao.queryBuilder().unique();
+                                if(userInfo!=null) {
+                                    Log.i(TAG, userInfo.toString());
+                                    if (!mid.equals(userInfo.getEid())) {
+                                        approveNumber.setId(mid);
+                                        approveNumberDao.insert(approveNumber);
+                                        back();
+                                    }else{
+                                        Toast.makeText(getApplicationContext(), "提示：审批人不得选自己，请选择其他审批人！", Toast.LENGTH_SHORT).show();
+                                    }
+                                }
+                            }
                         } else {
-                            Toast.makeText(getApplicationContext(), "提示：该审批人已被选择，请选择其他审批人！", Toast.LENGTH_SHORT).show();
+                            if(index!=4) {
+                                Toast.makeText(getApplicationContext(), "提示：该审批人已被选择，请选择其他审批人！", Toast.LENGTH_SHORT).show();
+                            }else{
+                                Toast.makeText(getApplicationContext(), "提示：该参会人已被选择，请选择其他参会人！", Toast.LENGTH_SHORT).show();
+                            }
                         }
                     }    //Log.i(TAG, "被点击的员工Eid：" + approveNumber.toString());
                     return false;
